Add symbols and move pause/resume button
