+++ src/api/Reducers/deltaTimes.js
@@ -3,6 +3,8 @@ import { actionTypes } from '../Actions/actionTypes';
 const defaultState = {
   hasNextDeltaTimeStep: undefined,
   hasPrevDeltaTimeStep: undefined,
+  nextDeltaTimeStep: undefined,
+  prevDeltaTimeStep: undefined
 };
 
 // TODO: this is where I should change how the data I send from the topic should be converted to states in the store
